+++ src/components/index.ts
@@ -3,5 +3,6 @@ export { GithubStats } from './GithubStats';
 export { ErrorPage404, ErrorPage500 } from './ErrorPage';
 
 export { MainPage } from './MainPage';
+export { IntroPage } from './IntroPage';
 export { ProjectsPage } from './ProjectsPage';
 export { ProjectPage } from './ProjectPage';
